refs #2 refactor com.airwallex.calculator.main.component to com.airwallex.calculator.component
