+++ src/main/java/com/airwallex/calculator/operator/action/RedoOperator.java
@@ -1,7 +1,7 @@
 package com.airwallex.calculator.operator.action;
 
-import com.airwallex.calculator.main.component.NumberStack;
-import com.airwallex.calculator.main.component.OperatorStack;
+import com.airwallex.calculator.component.NumberStack;
+import com.airwallex.calculator.component.OperatorStack;
 import com.airwallex.calculator.operator.Operator;
 import com.airwallex.calculator.operator.anotation.OperatorCommand;
