+++ src/test/java/com/airwallex/calculator/component/NumberStackTest.java
@@ -1,11 +1,11 @@
-package com.airwallex.calculator.main.component;
+package com.airwallex.calculator.component;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
-import com.airwallex.calculator.main.component.NumberStack;
-import com.airwallex.calculator.main.component.RealNumber;
+import com.airwallex.calculator.component.NumberStack;
+import com.airwallex.calculator.component.RealNumber;
 
 @SpringBootTest
 public class NumberStackTest {
