+++ src/test/java/com/airwallex/calculator/component/RealNumberTest.java
@@ -1,10 +1,10 @@
-package com.airwallex.calculator.main.component;
+package com.airwallex.calculator.component;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
-import com.airwallex.calculator.main.component.RealNumber;
+import com.airwallex.calculator.component.RealNumber;
 
 @SpringBootTest
 public class RealNumberTest {
