+++ src/test/java/com/airwallex/calculator/operator/action/ClearOperatorTest.java
@@ -5,8 +5,8 @@ import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 import org.springframework.boot.test.context.SpringBootTest;
 
-import com.airwallex.calculator.main.component.NumberStack;
-import com.airwallex.calculator.main.component.OperatorStack;
+import com.airwallex.calculator.component.NumberStack;
+import com.airwallex.calculator.component.OperatorStack;
 import com.airwallex.calculator.operator.action.ClearOperator;
 
 @SpringBootTest
